Make install script more robust
Using plain `cp` and `mkdir` will fail when using a non-standard umask
for either compiling Terminix or running the installation script. Fix that
by using the `install` command with explicit modes when necessary.

Additionally, ensure the desktop file cache and icon cache are properly
updated when either /usr or /usr/share is chosen as PREFIX.