Don't call query_info which might block
g_file_query_info is a sync API that will do IO and potentially
block if the network connection is slow (or dropped). If that
call blocks, it will freeze nautilus since the extension is called
from the main thread. Since we only use the resulting information
for the filename add a tooltip use the get_name() provided by
nautilus itself instead.