Only show VTE config warning dialog when opening preferences
Since Tilix is the default terminal emulator on a couple of
distributions, throwing a warning message into the user's face is not a
great initial experience.
So, those distributions would disable the warning message, which is also
not great, because it hides an issue.

This patch implements a middle-ground by only showing the configuration
error when the user opens the preferences dialog.