Bump minimum VTE version to 0.46
This allows us to get rid of some additional complexity due to version
checks. Most notably though, this fixes a module initializer dependency
issue for the URL regular expressions, which resolves #1996.
Since 0.46 or a higher version is in most stable distributions, this
change should not be an issue for most users (and we do check the VTE
version now at Tilix startup).