Add "Scale (fit)" background image mode
The current "Scale" mode crops out part of the image if the scaled
image is larger than the terminal in any direction. This is useful for
most cases, such as wallpapers repurposed as terminal backgrounds.
However, this may be surprising to users coming from xfce4-terminal,
where the equivalent "Scaled" mode tries to fit the background within
the terminal in both dimensions. This new "Scale (fit)" mode is meant
to emulate this behaviour.